Wrap checkmate::check_choice to handle its output differently

This wrapper always returns TRUE or FALSE so that it can be used in conditional statements. If the check fails, it will optionally emit the error message returned by checkmate.

Usage

chk_choice(..., quiet = TRUE)

Arguments

...

Arguments passed on to the specified checkmate function.

quiet

A logical which is TRUE if a failed check should not emit its error message; FALSE if it should. Default: TRUE

Value

A logical which is TRUE if the check succeeded and FALSE if not. The message generated by a failed check is set as an attribute of the return value.